requirements.txt
absl-py==0.15.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
clang==5.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
datasets==1.5.0
dill==0.3.6
filelock==3.12.0
flatbuffers==1.12
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==3.1.0
huggingface-hub==0.0.19
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
multidict==6.0.4
multiprocess==0.70.14
numpy==1.19.5
oauthlib==3.2.2
opencv-contrib-python==4.5.1.48
opt-einsum==3.3.0
packaging==23.1
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
pytz==2023.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
torchaudio==0.7.2
tqdm==4.49.0
typing-extensions==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.0.3
wrapt==1.12.1
xxhash==3.2.0
yarl==1.9.2
zipp==3.6.0
